requirements.txt
alembic==1.8.1
attrs==22.1.0
autopep8==2.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
diskcache==5.4.0
Fiona==1.8.22
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
future==0.18.2
geopandas==0.12.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.23.5
packaging==21.3
pandas==1.5.2
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==1.10.2
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
s2sphere==0.2.5
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sqlalchemy-get-or-create==0.1.5
tabulate==0.8.10
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
